+++ docs/CHANGELOG.md
@@ -9,6 +9,10 @@ icon: lucide/scroll-text
 
 * [poetry] Bump default `uv_build` bounds to `>=0.12.0,<0.13` ([#921](https://github.com/osprey-oss/migrate-to-uv/pull/921))
 
+### Bug fixes
+
+* Default to "project" as project name when not inferable ([#922](https://github.com/osprey-oss/migrate-to-uv/pull/922))
+
 ## 0.12.0 - 2026-04-08
 
 The project source code was moved under [a new GitHub organization](https://github.com/osprey-oss/migrate-to-uv). This release does not come with functional changes, but since the documentation has been moved to a [new URL](https://osprey-oss.github.io/migrate-to-uv/), it ensures that links are updated in the CLI and in PyPI metadata. See https://github.com/osprey-oss/migrate-to-uv/issues/823 for details about the move to an organization.